Used memoized version of comment and thread date to stop rerendering dates
